Crud de productos completo y pagina de inicio

+++ organic/src/main/java/com/example/organic/Seeder/DataSeeder.java
@@ -25,12 +25,14 @@ public class DataSeeder implements CommandLineRunner {
 
     @Override
     public void run(String... args) {
-        System.out.println("🚀 Seeder ejecutándose...");
-        categoriasRepository.deleteAll();
-        condicionesRepository.deleteAll();
-        tiposRepository.deleteAll();
+        if (categoriasRepository.count() >3 && condicionesRepository.count()>3 && tiposRepository.count()>3) {
+            categoriasRepository.deleteAll();
+            condicionesRepository.deleteAll();
+            tiposRepository.deleteAll();
+        }
 
-        // Inserta datos
+        if (categoriasRepository.count()==0 && condicionesRepository.count()== 0 && tiposRepository.count()==0) {
+
         categoriasRepository.save(new CategoriasEntity("Shampoo", "producto para el lavado del cabello"));
         categoriasRepository.save(new CategoriasEntity("Acondicionador", "producto para hidratar y suavizar el cabello"));
         categoriasRepository.save(new CategoriasEntity("Tratamiento", "producto para reparar el cabello"));
@@ -43,6 +45,7 @@ public class DataSeeder implements CommandLineRunner {
         tiposRepository.save(new TiposCabellosEntity("Ondulado", "Cabello con ligeras ondulaciones"));
         tiposRepository.save(new TiposCabellosEntity("Rizado", "Cabello muy ondulado y difícil de manejar"));
 
-        System.out.println("✅ Datos insertados con éxito");
+        System.out.println("Datos insertados con éxito");
+        }
     }
 }
